+++ packages/harpguru-core/src/types/index.ts
@@ -1,4 +1,3 @@
-import { TapGestureHandlerStateChangeEvent } from 'react-native-gesture-handler'
 import type {
   ApparatusIds,
   HarpStrata,
@@ -22,6 +21,11 @@ export enum MenuStates {
   NoMenu,
 }
 
+export enum MenuStashPosition {
+  Top,
+  Bottom,
+}
+
 export enum CellStates {
   TappedOn,
   TappedOff,
@@ -29,12 +33,31 @@ export enum CellStates {
   Off,
 }
 
+// The reason we need to identify the animation
+// types as 'Unsafe' and 'Safe' rather than 'Recoiling'
+// and 'NonRecoiling' or something is because the most
+// important thing that the name needs to communicate
+// to the developer is that if they choose the
+// 'Unsafe' animation type, there is a chance that
+// they will encounter a nasty unmount error.
+// This will occur if the callback that is being
+// passed to the hook which uses this enum is updating
+// the  harpface, particularly if it is creating a
+// harp with fewer rows.
+export enum TapAnimationTypes {
+  Safe,
+  Unsafe,
+}
+
 export type MenuProps = {
-  readonly hideMenu: boolean
-  readonly hideLabel: boolean
-  readonly openCloseTapHandler: (
-    arg0: TapGestureHandlerStateChangeEvent
-  ) => void
+  readonly isMenuStashed: boolean
+  readonly isLabelHidden: boolean
+  readonly stashPosition: MenuStashPosition
+  readonly openCloseMenu: () => void
+}
+
+export type ChildrenProps = {
+  readonly children: React.ReactNode
 }
 
 export type SetActiveHarpStrata = (arg0: HarpStrata) => void