+++ packages/harpguru-core/src/utils/get-scaled-menu-label-protrusion/index.ts
@@ -1,0 +1,1 @@
+export { getScaledMenuLabelProtrusion } from './get-scaled-menu-label-protrusion'