+++ packages/harpguru-core/CHANGELOG.md
@@ -19,11 +19,16 @@ and this project adheres to ~~[Semantic Versioning](https://semver.org/spec/v2.0
 
 - MAJOR: Buffer for harp cell activity updates
 - MINOR: Add animations to harp-cell toggleing
+- MINOR: Add animations to menu interactions
 
 ### Changed
 
 - MINOR: Memoisation added to HarpCell
 - MINOR: Default harp setup to popular configuration (major pent in 2nd position)
+
+### Removed
+
+- MAJOR: Option swipe selection
 
 ## [v2.1.0](https://github.com/js-jslog/harpguru/releases/tag/v3.2.0) - 2020-10-19